requirements.txt
Twisted==14.0.0
autobahn==0.8.9
beautifulsoup4==4.3.2
command-not-found==0.3
ipython==2.0.0
lxml==3.3.5
matplotlib==1.3.1
nose==1.3.3
numpy==1.8.1
pandas==0.13.1
py-stackexchange==1.1-4
pycurl==7.19.3
pygobject==3.12.0
pymongo==2.7
pyparsing==2.0.2
python-apt==0.9.3.5
python-dateutil==2.2
pytz==2014.2
six==1.6.1
stackpy==0.1
tornado==3.2.1
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
zope.interface==4.1.1
